change bitfield index to uint32

+++ internal/bitfield/bitfield.go
@@ -2,22 +2,22 @@ package bitfield
 
 type Bitfield []byte
 
-func (bf Bitfield) HasPiece(index int) bool {
+func (bf Bitfield) HasPiece(index uint32) bool {
 	byteIndex := index / 8
 	offset := index % 8
 
-	if byteIndex < 0 || byteIndex >= len(bf) {
+	if byteIndex < 0 || byteIndex >= uint32(len(bf)) {
 		return false
 	}
 
 	return bf[byteIndex]>>uint(7-offset)&1 != 0
 }
 
-func (bf Bitfield) SetPiece(index int) {
+func (bf Bitfield) SetPiece(index uint32) {
 	byteIndex := index / 8
 	offset := index % 8
 
-	if byteIndex < 0 || byteIndex >= len(bf) {
+	if byteIndex < 0 || byteIndex >= uint32(len(bf)) {
 		return
 	}
